requirements.txt
Flask==2.3.0
mysql-connector-python==8.0.33
Werkzeug==2.3.0
cryptography==41.0.0
requests==2.31.0
python-dotenv==1.0.0
